[feature] add animation in lottery
- add animation for lottery number created
- add `ui/lottery/LotteryNumberItem`
